+++ src/scenes/AcceleratorRouter/Overview/index.tsx
@@ -6,14 +6,13 @@ import { makeStyles } from '@mui/styles';
 import { Tabs } from '@terraware/web-components';
 
 import Page from 'src/components/Page';
-import Button from 'src/components/common/button/Button';
-import { APP_PATHS } from 'src/constants';
 import { useLocalization, useUser } from 'src/providers';
 import ParticipantsList from 'src/scenes/AcceleratorRouter/Participants/ParticipantsList';
 import strings from 'src/strings';
-import useDeviceInfo from 'src/utils/useDeviceInfo';
 import useQuery from 'src/utils/useQuery';
 import useStateLocation, { getLocation } from 'src/utils/useStateLocation';
+
+import CohortsTable from '../Cohorts/CohortsTable';
 
 const useStyles = makeStyles((theme: Theme) => ({
   tabs: {
@@ -43,7 +42,6 @@ const TabPlaceholder = ({ name }: { name: string }) => {
 };
 
 const OverviewView = () => {
-  const { isMobile } = useDeviceInfo();
   const { isAllowed } = useUser();
   const { activeLocale } = useLocalization();
   const history = useHistory();
@@ -53,10 +51,6 @@ const OverviewView = () => {
   const location = useStateLocation();
 
   const [activeTab, setActiveTab] = useState<string>(tab);
-
-  const goToNewCohort = () => {
-    history.push(APP_PATHS.ACCELERATOR_COHORTS_NEW);
-  };
 
   const onTabChange = useCallback(
     (newTab: string) => {
@@ -89,7 +83,7 @@ const OverviewView = () => {
       {
         id: 'cohorts',
         label: strings.COHORTS,
-        children: <TabPlaceholder name={strings.COHORTS} />,
+        children: <CohortsTable />,
       },
     ];
   }, [activeLocale, isAllowed]);
@@ -103,16 +97,7 @@ const OverviewView = () => {
   }, [tab, tabs]);
 
   return (
-    <Page
-      title={strings.OVERVIEW}
-      rightComponent={
-        isMobile ? (
-          <Button id='new-cohort' icon='plus' onClick={goToNewCohort} size='medium' />
-        ) : (
-          <Button id='new-cohort' label={strings.ADD_COHORT} icon='plus' onClick={goToNewCohort} size='medium' />
-        )
-      }
-    >
+    <Page title={strings.OVERVIEW}>
       <Box display='flex' flexDirection='column' flexGrow={1} className={classes.tabs}>
         <Tabs activeTab={activeTab} onTabChange={onTabChange} tabs={tabs} />
       </Box>
